Move the sync writer into logging
I'm trying to get rid of the unfortunately generic "utils" package, so I'm moving this into logging, which is essentially how it's used.
